refactor(search): group util functions
